Add files via upload
This project is a Power BI dashboard designed to analyze and visualize the sales performance of "Plant Co." for the years 2023 and 2024. The dashboard provides key insights into Year-to-Date (YTD) and Prior Year-to-Date (PYTD) sales, gross profit margin, country-wise performance, and year-over-year comparisons. It highlights areas of underperformance, tracks sales trends by product type and month, and segments account profitability based on gross profit percentage and sales value.